Refactor proposal views into subpackage
